separate bloom filter based pruner out
